services/python/requirements.txt
# gRPC
grpcio>=1.74.0 --only-binary=:all:
grpcio-tools>=1.74.0 --only-binary=:all:
grpcio-health-checking>=1.74.0
grpcio-reflection>=1.74.0
protobuf>=4.25.1 --only-binary=:all:

# Machine Learning
torch>=2.1.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.16.0 --index-url https://download.pytorch.org/whl/cpu
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image Generation
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1
Pillow>=9.0.0  # For image processing

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
